requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0
torchmetrics>=0.11.0
tqdm>=4.66.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.0.0
torchvision>=0.15.0